Metadex plugin for WordPress v0.6
v0.6 The plugin now allows a wordpress page to be constructed from the Float settings. The Float visualisation still needs work which will be the focus of the next phase of development. The naming conventions have been defined but not implemented and documentation has commenced. The formalisation of the code structure and naming conventions will then be applied to a Wordpress page template (for a sample theme) and an external HTML interface to ensure rigour in that structure and allow a range of developers to use the code.